Refactor(jira): Move the base test action into alert main project.
