+++ src/main/java/com/synopsys/integration/alert/channel/jira/common/JiraGlobalTestAction.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.issuetracker.jira.common;
+package com.synopsys.integration.alert.channel.jira.common;
 
 import com.synopsys.integration.alert.common.action.TestAction;
 import com.synopsys.integration.alert.common.message.model.MessageResult;
